+++ scripts/world/Africa.json
