+++ scripts/world/Asia.json
