+++ scripts/world/Europe.json
