+++ scripts/world/Oceania.json
