+++ server/__init__.py
@@ -148,6 +148,8 @@ def add_api_endpoints(app):
         ProjectsStatisticsQueriesUsernameAPI,
         ProjectsStatisticsQueriesPopularAPI,
     )
+    from server.api.projects.teams import ProjectsTeamsAPI
+    from server.api.projects.campaigns import ProjectsCampaignsAPI
     from server.api.projects.actions import (
         ProjectsActionsTransferAPI,
         ProjectsActionsMessageContributorsAPI,
@@ -202,30 +204,24 @@ def add_api_endpoints(app):
     from server.api.licenses.actions import LicensesActionsAcceptAPI
 
     # Campaigns API endpoint
-    from server.api.campaigns.resources import (
-        CampaignAPI,
-        GetAllCampaignsAPI,
-        CampaignProjectAPI,
-        DeleteAllProjectCampaignsAPI,
-        CampaignOrganisationAPI,
-    )
+    from server.api.campaigns.resources import CampaignsRestAPI, CampaignsAllAPI
 
     # Organisations API endpoint
     from server.api.organisations.resources import (
-        OrganisationAPI,
-        GetAllOrganisationsAPI,
+        OrganisationsRestAPI,
+        OrganisationsAllAPI,
     )
+    from server.api.organisations.campaigns import OrganisationsCampaignsAPI
 
     # Countries API endpoint
     from server.api.countries.resources import CountriesRestAPI
 
     # Teams API endpoint
-    from server.api.teams.resources import (
-        TeamAPI,
-        ListTeamsAPI,
-        TeamMembersAPI,
-        TeamProjectsAPI,
-        DeleteMultipleTeamMembersAPI,
+    from server.api.teams.resources import TeamsRestAPI, TeamsAllAPI
+    from server.api.teams.actions import (
+        TeamsActionsJoinAPI,
+        TeamsActionsLeaveAPI,
+        TeamsActionsLeaveMultipleAPI,
     )
 
     # Notifications API endpoint
@@ -324,11 +320,33 @@ def add_api_endpoints(app):
         ProjectsStatisticsQueriesUsernameAPI,
         "/api/v2/projects/<int:project_id>/statistics/queries/<string:username>/",
     )
-
     api.add_resource(
         ProjectsStatisticsQueriesPopularAPI, "/api/v2/projects/queries/popular/"
     )
 
+    api.add_resource(
+        ProjectsTeamsAPI,
+        "/api/v2/projects/<int:project_id>/teams",
+        endpoint="get_all_project_teams",
+        methods=["GET"],
+    )
+    api.add_resource(
+        ProjectsTeamsAPI,
+        "/api/v2/projects/<int:project_id>/teams/<int:team_id>",
+        methods=["PUT", "DELETE", "PATCH"],
+    )
+    api.add_resource(
+        ProjectsCampaignsAPI,
+        "/api/v2/projects/<int:project_id>/campaigns",
+        endpoint="get_all_project_campaigns",
+        methods=["GET"],
+    )
+    api.add_resource(
+        ProjectsCampaignsAPI,
+        "/api/v2/projects/<int:project_id>/campaigns/<int:campaign_id>",
+        endpoint="assign_remove_campaign_to_project",
+        methods=["PUT", "DELETE"],
+    )
     # Projects actions endoints
     api.add_resource(
         ProjectsActionsMessageContributorsAPI,
@@ -344,6 +362,7 @@ def add_api_endpoints(app):
     api.add_resource(
         ProjectsActionsUnFeatureAPI,
         "/api/v2/projects/<int:project_id>/actions/remove-feature",
+        methods=["POST"],
     )
 
     api.add_resource(
@@ -489,14 +508,89 @@ def add_api_endpoints(app):
         "/api/v2/licenses/<int:license_id>/actions/accept-for-me/",
     )
 
-    # Campaigns REST endpoints
-    api.add_resource(CampaignsRestAPI, "/api/v2/campaigns/")
-
-    # Organisations REST endpoints
-    api.add_resource(OrganisationsRestAPI, "/api/v2/organisations/")
-
     # Countries REST endpoints
     api.add_resource(CountriesRestAPI, "/api/v2/countries/")
+
+    # Organisations REST endpoints
+    api.add_resource(OrganisationsAllAPI, "/api/v2/organisations", methods=["GET"])
+    api.add_resource(
+        OrganisationsAllAPI,
+        "/api/v2/organisations",
+        endpoint="create_organisation",
+        methods=["POST"],
+    )
+    api.add_resource(
+        OrganisationsRestAPI,
+        "/api/v2/organisations/<int:organisation_id>",
+        endpoint="get_organisation",
+        methods=["GET"],
+    )
+    api.add_resource(
+        OrganisationsRestAPI,
+        "/api/v2/organisations/<int:organisation_id>",
+        methods=["PUT", "DELETE"],
+    )
+
+    # Organisations additional resources endpoints
+    api.add_resource(
+        OrganisationsCampaignsAPI,
+        "/api/v2/organisations/<int:organisation_id>/campaigns",
+        endpoint="get_all_organisation_campaigns",
+        methods=["GET"],
+    )
+    api.add_resource(
+        OrganisationsCampaignsAPI,
+        "/api/v2/organisations/<int:organisation_id>/campaigns/<int:campaign_id>",
+        endpoint="assign_campaign_to_organisation",
+        methods=["PUT", "DELETE"],
+    )
+
+    # Teams REST endpoints
+    api.add_resource(TeamsAllAPI, "/api/v2/teams", methods=["GET"])
+    api.add_resource(
+        TeamsAllAPI, "/api/v2/teams", endpoint="create_team", methods=["POST"]
+    )
+    api.add_resource(
+        TeamsRestAPI, "/api/v2/teams/<int:team_id>", methods=["GET", "PUT", "DELETE"]
+    )
+
+    # Teams actions endpoints
+    api.add_resource(
+        TeamsActionsJoinAPI,
+        "/api/v2/teams/<int:team_id>/actions/join",
+        methods=["POST"],
+    )
+    api.add_resource(
+        TeamsActionsLeaveAPI,
+        "/api/v2/teams/<int:team_id>/actions/leave",
+        endpoint="leave_team",
+        methods=["POST"],
+    )
+    api.add_resource(
+        TeamsActionsLeaveMultipleAPI,
+        "/api/v2/teams/<int:team_id>/actions/remove-users",
+        endpoint="remove_users_from_team",
+        methods=["POST"],
+    )
+
+    # Campaigns REST endpoints
+    api.add_resource(
+        CampaignsAllAPI,
+        "/api/v2/campaigns",
+        endpoint="get_all_campaign",
+        methods=["GET"],
+    )
+    api.add_resource(
+        CampaignsAllAPI,
+        "/api/v2/campaigns",
+        endpoint="create_campaign",
+        methods=["POST"],
+    )
+    api.add_resource(
+        CampaignsRestAPI,
+        "/api/v2/campaigns/<int:campaign_id>",
+        methods=["GET", "PUT", "DELETE"],
+    )
 
     # Notifications REST endpoints
     api.add_resource(NotificationsRestAPI, "/api/v2/notifications/<int:message_id>/")
@@ -581,97 +675,4 @@ def add_api_endpoints(app):
         "/api/v2/system/authentication/applications/<string:application_key>/",
         endpoint="check_application",
         methods=["PATCH"],
-    )
-    api.add_resource(
-        OrganisationAPI,
-        "/api/v1/organisations",
-        endpoint="create_organisation",
-        methods=["POST"],
-    )
-    api.add_resource(GetAllOrganisationsAPI, "/api/v1/organisations", methods=["GET"])
-    api.add_resource(
-        OrganisationAPI,
-        "/api/v1/organisations/<int:organisation_id>",
-        endpoint="get_organisation",
-        methods=["GET"],
-    )
-    api.add_resource(
-        OrganisationAPI,
-        "/api/v1/organisations/<int:organisation_id>",
-        methods=["PUT", "DELETE"],
-    )
-    api.add_resource(
-        CampaignOrganisationAPI,
-        "/api/v1/organisations/<int:organisation_id>/campaigns",
-        endpoint="get_all_organisation_campaigns",
-        methods=["GET"],
-    )
-    api.add_resource(
-        CampaignOrganisationAPI,
-        "/api/v1/organisations/<int:organisation_id>/campaigns/<int:campaign_id>",
-        endpoint="assign_campaign_to_organisation",
-        methods=["PUT", "DELETE"],
-    )
-    api.add_resource(TeamAPI, "/api/v1/teams", endpoint="create_team", methods=["POST"])
-    api.add_resource(ListTeamsAPI, "/api/v1/teams", methods=["GET"])
-    api.add_resource(
-        TeamAPI, "/api/v1/teams/<int:team_id>", methods=["GET", "PUT", "DELETE"]
-    )
-    api.add_resource(
-        TeamMembersAPI, "/api/v1/teams/<int:team_id>/actions/join", methods=["POST"]
-    )
-    api.add_resource(
-        TeamMembersAPI,
-        "/api/v1/teams/<int:team_id>/actions/leave",
-        endpoint="leave_team",
-        methods=["DELETE"],
-    )
-    api.add_resource(
-        DeleteMultipleTeamMembersAPI,
-        "/api/v1/teams/<int:team_id>/actions/remove-users",
-        endpoint="remove_users_from_team",
-        methods=["DELETE"],
-    )
-    api.add_resource(
-        TeamProjectsAPI,
-        "/api/v1/projects/<int:project_id>/teams",
-        endpoint="get_all_project_teams",
-        methods=["GET"],
-    )
-    api.add_resource(
-        TeamProjectsAPI,
-        "/api/v1/projects/<int:project_id>/teams/<int:team_id>",
-        methods=["PUT", "DELETE", "PATCH"],
-    )
-    api.add_resource(
-        CampaignAPI, "/api/v1/campaigns", endpoint="create_campaign", methods=["POST"]
-    )
-    api.add_resource(
-        GetAllCampaignsAPI,
-        "/api/v1/campaigns",
-        endpoint="get_all_campaign",
-        methods=["GET"],
-    )
-    api.add_resource(
-        CampaignAPI,
-        "/api/v1/campaigns/<int:campaign_id>",
-        methods=["GET", "PUT", "DELETE"],
-    )
-    api.add_resource(
-        CampaignProjectAPI,
-        "/api/v1/projects/<int:project_id>/campaigns",
-        endpoint="get_all_project_campaigns",
-        methods=["GET"],
-    )
-    api.add_resource(
-        CampaignProjectAPI,
-        "/api/v1/projects/<int:project_id>/campaigns/<int:campaign_id>",
-        endpoint="assign_remove_campaign_to_project",
-        methods=["PUT", "DELETE"],
-    )
-    api.add_resource(
-        DeleteAllProjectCampaignsAPI,
-        "/api/v1/projects/<int:project_id>/campaigns/actions/delete-all",
-        endpoint="delete_all_project_campaigns",
-        methods=["POST"],
     )
